Changed ints to uint_8 to decrease data an program bytes

+++ 171114-103337-uno/src/ArduinoNunchuk.cpp
@@ -24,8 +24,8 @@ void ArduinoNunchuk::init()
 //functie om te controleren of er iets op de nunchuck veranderd is.
 void ArduinoNunchuk::update()
 {
-  int count = 0;
-  int values[6];
+  uint8_t count = 0;
+  uint8_t values[6];
 
   //vraag gegevens op
   Wire.requestFrom(ADDRESS, 6);
